Fix Pr comment
