+++ src/views/RegionProfile/i18n.json
@@ -4,6 +4,7 @@
         "regionalProfileNationalSocietyName": "National Society in",
         "regionalProfileFundingAndOperationsDashboards": "Funding and Operations Dashboards",
         "regionalProfileAdditionalLinks": "Additional Links",
-        "regionProfileContacts": "Contacts"
+        "regionProfileContacts": "Contacts",
+        "regionalProfileNationalSocietyTitle": "National Society in"
     }
 }
